+++ apps/user/src/app/form/page.tsx
@@ -6,7 +6,7 @@ import 출신학교및학력 from './출신학교및학력/출신학교및학력
 import 전형선택 from './전형선택/전형선택';
 import 성적입력 from './성적입력/성적입력';
 import 자기소개서 from './자기소개서/자기소개서';
-import 완료 from './완료/완료';
+import 초안제출 from './초안제출/초안제출';
 import 최종제출 from './최종제출/최종제출';
 import { useFormStepState } from '@/hooks/state/useFormStepState';
 import { SwitchCase } from '@toss/react';
@@ -24,7 +24,7 @@ const FormPage = () => {
                 전형선택: <전형선택 />,
                 성적입력: <성적입력 />,
                 자기소개서: <자기소개서 />,
-                완료: <완료 />,
+                초안제출: <초안제출 />,
                 최종제출: <최종제출 />,
             }}
             defaultComponent={<지원자정보 />}
