+++ apps/user/src/app/form/초안제출/초안제출.tsx
@@ -5,15 +5,15 @@ import { AppLayout } from '@/layouts';
 import { useInterval } from '@maru/hooks';
 import { IconCancelCircle, IconCheckCircle } from '@maru/icon';
 import { color, font } from '@maru/theme';
-import { Button, Column, Row } from '@maru/ui';
+import { Button, Column, Row, Text } from '@maru/ui';
 import { flex } from '@maru/utils';
 import { useState } from 'react';
-import { useCheckFilledForm, useCTAButton } from './완료.hooks';
+import { useCheckFilledForm, useCTAButton } from './초안제출.hooks';
 import { CompleteAlaram, CheckFormComplete, DraftFormConfirm } from '@/components/form';
 import styled from 'styled-components';
 import { useOverlay } from '@toss/use-overlay';
 
-const 완료 = () => {
+const 초안제출 = () => {
     const overlay = useOverlay();
     const [isShowCompleteAlaram, setIsShowCompleteAlaram] = useState(true);
     const { setFormStep } = useFormStepState();
@@ -40,7 +40,7 @@ const 완료 = () => {
             {isShowCompleteAlaram ? (
                 <CompleteAlaram isFilledForm={isFilledForm} />
             ) : (
-                <Styled완료>
+                <Styled초안제출>
                     <Row gap={8} style={{ marginBottom: '55px' }} alignItems="center">
                         <CircleIconBox>
                             {isFilledForm ? (
@@ -49,41 +49,42 @@ const 완료 = () => {
                                 <IconCancelCircle width="100%" height="100%" />
                             )}
                         </CircleIconBox>
-                        <AlertMessage>
+                        <Text fontType="H1" color={color.gray900}>
                             {isFilledForm
                                 ? '원서 초안 작성 완료'
                                 : '아직 작성하지 않은 곳이 있어요'}
-                        </AlertMessage>
+                        </Text>
                     </Row>
                     <Column gap={12}>
-                        <Desc>
-                            {isFilledForm ? (
-                                <p>원서 접수에 필요한 초안을 모두 작성하셨습니다.</p>
-                            ) : (
-                                <p>원서 작성 중 입력하지 않은 곳이 있습니다.</p>
-                            )}
-                        </Desc>
-                        <StressDesc>
-                            {isFilledForm ? (
-                                <p>
-                                    원서 초안 제출 시 부산소프트웨어마이스터고등학교 입학전형에
-                                    응시한 것으로 처리되며
-                                    <br />더 이상 입학원서 수정이 불가능합니다.
-                                </p>
-                            ) : (
-                                <p>
-                                    원서 작성 중 입력하지 않은 곳이 있다면 초안을 제출할 수
-                                    없습니다.
-                                </p>
-                            )}
-                        </StressDesc>
-                        <Desc>
-                            {isFilledForm ? (
-                                <p>잘못 입력한 곳이 있는지 면밀히 검토해주시기 바랍니다.</p>
-                            ) : (
-                                <p>또한 잘못 입력한 곳이 있는지 면밀히 검토해주시기 바랍니다.</p>
-                            )}
-                        </Desc>
+                        {isFilledForm ? (
+                            <Text fontType="p1" color={color.gray900}>
+                                원서 접수에 필요한 초안을 모두 작성하셨습니다.
+                            </Text>
+                        ) : (
+                            <Text fontType="p1" color={color.gray900}>
+                                원서 작성 중 입력하지 않은 곳이 있습니다.
+                            </Text>
+                        )}
+                        {isFilledForm ? (
+                            <Text fontType="p1" color={color.red}>
+                                원서 초안 제출 시 부산소프트웨어마이스터고등학교 입학전형에 응시한
+                                것으로 처리되며
+                                <br />더 이상 입학원서 수정이 불가능합니다.
+                            </Text>
+                        ) : (
+                            <Text fontType="p1" color={color.red}>
+                                원서 작성 중 입력하지 않은 곳이 있다면 초안을 제출할 수 없습니다.
+                            </Text>
+                        )}
+                        {isFilledForm ? (
+                            <Text fontType="p1" color={color.gray900}>
+                                잘못 입력한 곳이 있는지 면밀히 검토해주시기 바랍니다.
+                            </Text>
+                        ) : (
+                            <Text fontType="p1" color={color.gray900}>
+                                또한 잘못 입력한 곳이 있는지 면밀히 검토해주시기 바랍니다.
+                            </Text>
+                        )}
                     </Column>
                     <CheckFormCompleteBox>
                         <CheckFormComplete
@@ -126,7 +127,9 @@ const 완료 = () => {
 
                     {isFilledForm && (
                         <Column gap={24}>
-                            <Question>제출하시겠습니까?</Question>
+                            <Text fontType="H3" color={color.gray900}>
+                                제출하시겠습니까?
+                            </Text>
                             <Row gap={16}>
                                 <Button
                                     onClick={handleAgainCheckFormButtonClick}
@@ -140,15 +143,15 @@ const 완료 = () => {
                             </Row>
                         </Column>
                     )}
-                </Styled완료>
+                </Styled초안제출>
             )}
         </AppLayout>
     );
 };
 
-export default 완료;
+export default 초안제출;
 
-const Styled완료 = styled.div`
+const Styled초안제출 = styled.div`
     ${flex({ flexDirection: 'column' })}
     max-width: 800px;
     height: 100%;
@@ -160,21 +163,6 @@ const CircleIconBox = styled.div`
     height: 64px;
 `;
 
-const AlertMessage = styled.p`
-    ${font.H1}
-    color: ${color.gray900};
-`;
-
-const Desc = styled.p`
-    ${font.p1}
-    color: ${color.gray900};
-`;
-
-const StressDesc = styled.p`
-    ${font.H4}
-    color: ${color.red};
-`;
-
 const CheckFormCompleteBox = styled.div`
     display: flex;
     flex-wrap: wrap;
@@ -182,9 +170,4 @@ const CheckFormCompleteBox = styled.div`
     width: 100%;
     height: 248px;
     margin: 48px 0px;
-`;
-
-const Question = styled.p`
-    ${font.H3}
-    color: ${color.black};
 `;
