+++ apps/user/src/components/form/index.ts
@@ -7,4 +7,4 @@ export { default as ProgressBar } from './ProgressBar/ProgressBar';
 export { default as CompleteAlaram } from './CompleteAlaram/CompleteAlaram';
 export { default as CheckFormComplete } from './CheckFormComplete/CheckFormComplete';
 export { default as DraftFormConfirm } from './DraftFormConfirm/DraftFormConfirm';
-export { default as 최종제출Table } from './최종제출Table/최종제출Table';
+export { default as FinalFormTable } from './FinalFormTable/FinalFormTable';
